inventory

+++ app/Http/Controllers/Panel/FactorController.php
@@ -10,6 +10,7 @@ use App\Models\Invoice;
 use App\Models\Product;
 use App\Models\Province;
 use Illuminate\Http\Request;
+use function Symfony\Component\String\b;
 
 class FactorController extends Controller
 {
@@ -53,6 +54,11 @@ class FactorController extends Controller
         // edit own invoice OR is admin
         $this->authorize('edit-invoice', $factor->invoice);
 
+        // if this factor has output from inventory redirect it
+        if ($factor->inventory_report){
+            return back();
+        }
+
         return view('panel.factors.edit', compact('factor'));
     }
 
@@ -63,6 +69,11 @@ class FactorController extends Controller
 
         // edit own invoice OR is admin
         $this->authorize('edit-invoice', $factor->invoice);
+
+        // if this factor has output from inventory redirect it
+        if ($factor->inventory_report){
+            return back();
+        }
 
         $invoice = $factor->invoice;
 
@@ -99,6 +110,11 @@ class FactorController extends Controller
     public function destroy(Factor $factor)
     {
         $this->authorize('invoices-delete');
+
+        // if this factor has output from inventory redirect it
+        if ($factor->inventory_report){
+            return back();
+        }
 
         $invoice = Invoice::find($factor->invoice_id);
         $factor->delete();
